requirements.txt
gin_config==0.5.0
lightning==2.5.5
matplotlib==3.10.7
numpy==2.3.3
pandas==2.3.3
Pillow==11.3.0
ray==2.49.2
rdkit==2025.3.6
scikit_learn==1.7.2
scipy==1.16.2
torch==2.8.0
xgboost==3.0.5
